add package.json engines and corresponding .nvmrc
I wish the nvmrc were not necessary, but it still is.
See https://github.com/nvm-sh/nvm/issues/651

I think nvmrc is the most popular Node version manager, so it _probably_
makes sense to add the `.nvmrc`. If we want to skip this and go just
with the `package.json`, that's fine too.