docs: better formatting for README
This improves the Markdown formatting in multiple ways. Most notably, it
nests things correctly under numbered lists, and makes the whole thing
more readable _as markdown_, rather than needing to be rendered to HTML
to be easily-readable (that's the point of Markdown, after all).

It also makes it more clear that there are two options for running the
backend, and double-highlights that you need to run `make build-docker`,
because I missed that for way too long personally 🙃

This also increases most heading sizes and uses the underline syntax for
h1 & h2, again improving readability as both markdown & html versions.