feat: no more embedding secret keys
This copies ideas from the contract in
https://github.com/AhaLabs/soroban-abundance-token, allowing anyone to
mint themselves tokens. Now this can be used from the frontend directly,
so we no longer need to embed the admin's secret key in the JS. Hooray!

Upon closer inspection, we didn't need any of those constants anymore,
since the contract IDs are embedded in the token libraries.